updated xml files so that margins and textsizes would be conformed across the app
