+++ FoodFriends/app/src/main/java/com/example/foodfriends/models/UserLike.java
@@ -6,8 +6,8 @@ import com.parse.ParseUser;
 
 @ParseClassName("UserLike")
 public class UserLike extends ParseObject {
-    private static final String USER_KEY = "user";
-    private static final String RESTAURANT_KEY = "restaurant";
+    public static final String USER_KEY = "user";
+    public static final String RESTAURANT_KEY = "restaurant";
 
     public ParseUser getUser(){
         return getParseUser(USER_KEY);
