+++ FoodFriends/app/src/main/res/layout/fragment_profile.xml
@@ -10,7 +10,7 @@
         android:id="@+id/ivProfilePfp"
         android:layout_width="81dp"
         android:layout_height="81dp"
-        android:layout_marginTop="30dp"
+        android:layout_marginTop="@dimen/margin_padding_size_xlarge"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
@@ -20,9 +20,9 @@
         android:id="@+id/tvprofileUsername"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginTop="20dp"
+        android:layout_marginTop="@dimen/margin_padding_size_large"
         android:text="TextView"
-        android:textColor="#000000"
+        android:textColor="@color/black"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.498"
         app:layout_constraintStart_toStartOf="parent"
@@ -33,7 +33,7 @@
         android:id="@+id/rvProfilePosts"
         android:layout_width="414dp"
         android:layout_height="0dp"
-        android:layout_marginBottom="4dp"
+        android:layout_marginBottom="@dimen/margin_padding_size_tiny"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.666"
@@ -44,7 +44,7 @@
         android:id="@+id/ivAddPfp"
         android:layout_width="40dp"
         android:layout_height="48dp"
-        android:layout_marginTop="16dp"
+        android:layout_marginTop="@dimen/margin_padding_size_medium"
         android:clickable="true"
         android:src="@drawable/ic_baseline_add_a_photo_24"
         app:layout_constraintEnd_toEndOf="parent"
@@ -56,7 +56,7 @@
         android:id="@+id/profileTab"
         android:layout_width="0dp"
         android:layout_height="60dp"
-        android:layout_marginTop="11dp"
+        android:layout_marginTop="@dimen/margin_padding_size_small"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/tvprofileUsername"
@@ -81,8 +81,8 @@
         android:id="@+id/imageView"
         android:layout_width="60dp"
         android:layout_height="60dp"
-        android:layout_marginStart="44dp"
-        android:layout_marginTop="48dp"
+        android:layout_marginStart="@dimen/margin_padding_size_xxlarge"
+        android:layout_marginTop="@dimen/margin_padding_size_xxlarge"
         android:src="@drawable/ic_baseline_person_add_24"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
@@ -92,9 +92,9 @@
         android:id="@+id/btnLogOut"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginTop="16dp"
-        android:layout_marginEnd="16dp"
-        android:backgroundTint="#F89880"
+        android:layout_marginTop="@dimen/margin_padding_size_medium"
+        android:layout_marginEnd="@dimen/margin_padding_size_medium"
+        android:backgroundTint="@color/coral"
         android:text="LOG OUT"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
