+++ FoodFriends/app/src/main/res/layout/item_restaurant.xml
@@ -83,7 +83,7 @@
         android:id="@+id/imageView4"
         android:layout_width="43dp"
         android:layout_height="38dp"
-        android:layout_marginTop="8dp"
+        android:layout_marginTop="@dimen/margin_padding_size_small"
         app:layout_constraintEnd_toStartOf="@+id/tvExploreAddress"
         app:layout_constraintHorizontal_bias="0.533"
         app:layout_constraintStart_toStartOf="parent"
